pillar: move runInDebugContainer to containerd pkg
so that it cannot only be used by bpftrace, but also by
the collectinfo to run collect-info.sh from LOC
